pyproject.toml
[project]
name = "windsurf-throttle"
version = "0.1.0"
description = "Streamlit app for managing Windsurf credit caps"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your.email@example.com" }]
keywords = ["windsurf", "credit", "throttle", "streamlit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
    "pandas>=2.0.0",
    "watchdog>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/PFPT-Internal/windsurf-throttle"
Repository = "https://github.com/PFPT-Internal/windsurf-throttle"

[project.scripts]
windsurf-throttle = "windsurf_throttle.cli:main"

[project.gui-scripts]
windsurf-throttle-gui = "windsurf_throttle.app:run"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/windsurf_throttle"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
